requirements.txt
asgiref==3.6.0
astroid==2.12.14
colorama==0.4.6
dill==0.3.6
Django==4.1.5
djangorestframework==3.14.0
isort==5.11.4
lazy-object-proxy==1.9.0
mccabe==0.7.0
platformdirs==2.6.2
pylint==2.15.9
pylint-django==2.5.3
pylint-plugin-utils==0.7
pytz==2022.7
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
typing-extensions==4.4.0
tzdata==2022.7
wrapt==1.14.1
